requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Cartopy==0.23.0
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.1.1
django-appconf==1.0.6
django-compressor==4.5.1
django-imagekit==5.0.0
django-taggit==6.0.0
django_csp==3.8
fonttools==4.53.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pilkit==3.0
pillow==10.3.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==1.3.1
rjsmin==1.2.2
shapely==2.0.4
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tweepy==4.14.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
